update the document for the taskflow and add the taskflow of knowledge_minin
