+++ paddlenlp/taskflow/lexical_analysis.py
@@ -45,6 +45,7 @@ usage = r"""
            '''
            [{'text': 'LAC是个优秀的分词工具', 'segs': ['LAC', '是', '个', '优秀', '的', '分词', '工具'], 'tags': ['nz', 'v', 'q', 'a', 'u', 'n', 'n']}]
            '''
+
            lac(["LAC是个优秀的分词工具", "三亚是一个美丽的城市"])
            '''
            [{'text': 'LAC是个优秀的分词工具', 'segs': ['LAC', '是', '个', '优秀', '的', '分词', '工具'], 'tags': ['nz', 'v', 'q', 'a', 'u', 'n', 'n']},
